Initial Campaign and Product repositories

+++ src/main/java/com/maor/onlinesponsoredads/model/Product.java
@@ -26,7 +26,7 @@ public class Product {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer serialNumber;
+  private Long serialNumber;
 
   @ManyToMany(mappedBy = "products")
   private List<Campaign> campaigns;
